+++ src/rust/api.h
@@ -1,1 +1,7 @@
-SEXP savvy_clarabel_solve__ffi(SEXP m, SEXP n, SEXP Ai, SEXP Ap, SEXP Ax, SEXP b, SEXP q, SEXP Pi, SEXP Pp, SEXP Px, SEXP cone_spec, SEXP r_settings);
+SEXP savvy_clarabel_solve__ffi(SEXP c_arg__m, SEXP c_arg__n, SEXP c_arg__Ai, SEXP c_arg__Ap, SEXP c_arg__Ax, SEXP c_arg__b, SEXP c_arg__q, SEXP c_arg__Pi, SEXP c_arg__Pp, SEXP c_arg__Px, SEXP c_arg__cone_spec, SEXP c_arg__r_settings);
+
+// methods and associated functions for ClarabelSolver
+SEXP savvy_ClarabelSolver_is_update_allowed__ffi(SEXP self__);
+SEXP savvy_ClarabelSolver_new__ffi(SEXP c_arg__m, SEXP c_arg__n, SEXP c_arg__Ai, SEXP c_arg__Ap, SEXP c_arg__Ax, SEXP c_arg__b, SEXP c_arg__q, SEXP c_arg__Pi, SEXP c_arg__Pp, SEXP c_arg__Px, SEXP c_arg__cone_spec, SEXP c_arg__r_settings);
+SEXP savvy_ClarabelSolver_solve__ffi(SEXP self__);
+SEXP savvy_ClarabelSolver_update_data__ffi(SEXP self__, SEXP c_arg__Px, SEXP c_arg__Ax, SEXP c_arg__q, SEXP c_arg__b);